modify vo and entity

+++ src/main/java/xmu/crms/view/vo/GroupDetailsVO.java
@@ -1,9 +1,10 @@
 package xmu.crms.view.vo;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class GroupDetailsVO {
-    private int id;
+    private BigInteger id;
     private String name;
     private UserVO leader;
     private List<UserVO> members;
@@ -19,7 +20,7 @@ public class GroupDetailsVO {
         this.report = report;
     }
 
-    public GroupDetailsVO(int id, String name, UserVO leader, List<UserVO> members, List<TopicVO> topics, String report) {
+    public GroupDetailsVO(BigInteger id, String name, UserVO leader, List<UserVO> members, List<TopicVO> topics, String report) {
         this.id = id;
         this.name = name;
         this.leader = leader;
@@ -28,11 +29,11 @@ public class GroupDetailsVO {
         this.report = report;
     }
 
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
